% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BKMRfits10.R
\docType{data}
\name{BKMRfits10}
\alias{BKMRfits10}
\title{Sample Data for MI BKMR}
\format{
An object of class \code{list} of length 10.
}
\usage{
BKMRfits10
}
\description{
A sample data that is for MI BKMR illustration

\describe{
\item{BKMRfits10}{A list of 20 Multiple imputed datasets}
\item{BKMRfits10[[1]]}{An object contatinint the results return by the kmbayes function}
}
}
\keyword{datasets}
