% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDE.bkmr.MI.R
\name{CDE.bkmr.MI}
\alias{CDE.bkmr.MI}
\title{Estimate controlled direct effect for BKMR-CMA-MI}
\usage{
CDE.bkmr.MI(
  a,
  astar,
  e.y,
  m.value = NULL,
  m.quant = c(0.1, 0.5, 0.75),
  BKMRfits.Y,
  alpha = 0.05,
  sel,
  seed
)
}
\arguments{
\item{a}{exposure variables at current level}

\item{astar}{exposure variables at counterfactual level}

\item{e.y}{effect modifier for the outcome variable}

\item{m.value}{values that the mediator is set to}

\item{m.quant}{values of the quantile that the mediator is set to}

\item{BKMRfits.Y}{A list of `K` model fit regressing outcome on exposures, effect modifiers, mediator and confounders on outcome}

\item{alpha}{1-confidence interval}

\item{sel}{a vector selecting which iterations of thfe fit should be retained or inference}

\item{seed}{the random seed to use to evaluate the code}
}
\value{
Controlled direct effect for BKMR
}
\description{
Estimate controlled direct effect for BKMR-CMA-MI
}
\details{
For guided examples, go to
https://zc2326.github.io/causalbkmr/articles/BKMRCMA_QuickStart.html
}
