% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDERiskSummaries.CMA.R
\name{CDERiskSummaries.CMA}
\alias{CDERiskSummaries.CMA}
\title{CDERiskSummaries.CMA}
\usage{
CDERiskSummaries.CMA(
  fit.y,
  e.y = NULL,
  e.y.names = NULL,
  m.value = NULL,
  m.quant = c(0.1, 0.5, 0.75),
  m.name,
  qs = seq(0.25, 0.75, by = 0.05),
  q.fixed = 0.5,
  alpha = 0.05,
  sel,
  seed = 122
)
}
\arguments{
\item{fit.y}{An object contatinint the results return by the kmbayes function, a model fit regressing outcome on exposures, effect modifiers, mediator and confounders on outcome}

\item{e.y}{effect modifier for the outcome variable}

\item{e.y.names}{column name of the effect modifier for the outcome variable}

\item{m.value}{values that the mediator is set to}

\item{m.quant}{values of the quantile that the mediator is set to}

\item{m.name}{column name of the mediator}

\item{qs}{vector of quantiles at which to calculate the overall risk summary}

\item{q.fixed}{a second quantile at which to compare the estimated h function}

\item{alpha}{1-confidence interval}

\item{sel}{a vector selecting which iterations of the fit should be retained or inference}

\item{seed}{the random seed to use to evaluate the code}
}
\value{
a data frame containing the (posterior mean) estimate and posterior standard deviation of the CDE risk measures
}
\description{
Compare estimated Controlled Direct Effect when all predictors are at a particular quantile to when all are at a second fixed quantile
}
\details{
For guided examples, go to
https://zc2326.github.io/causalbkmr/articles/BKMRCMA_QuickStart.html
}
\examples{

\dontrun{
CDEriskSummary10 = CDERiskSummaries.CMA(fit.y = fit.y, e.y = e.y10, e.y.name = "E.Y", m.name = "m", sel = sel)
}
}
