% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma_sampledata.R
\name{cma_sampledata}
\alias{cma_sampledata}
\title{CMA Sample Data Simulation}
\usage{
cma_sampledata(N, L, P, scenario, seed)
}
\arguments{
\item{N}{sample size of dataset}

\item{L}{number of mixture components}

\item{P}{number of covariates}

\item{scenario}{Scenarios  that are similar to the scenarios presented in the paper}

\item{seed}{the random seed to use to evaluate the code}
}
\value{
a data frame
}
\description{
Simulate simple sample dataset for BKMRCMA illustration, function to generate simulated data with multiple exposures and true mediated effect
}
