% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePostmeanHnew.exact.MI.R
\name{ComputePostmeanHnew.exact.MI}
\alias{ComputePostmeanHnew.exact.MI}
\title{Compute the posterior mean and variance of \code{h} at a new predictor values using exact method for MI BKMR fits}
\usage{
ComputePostmeanHnew.exact.MI(
  fit,
  y = NULL,
  Z = NULL,
  X = NULL,
  Znew = NULL,
  sel = NULL
)
}
\arguments{
\item{fit}{An object contatinint the results return by the kmbayes function}

\item{y}{a vector of outcome data of length \code{n}.}

\item{Z}{an \code{n-by-M} matrix of predictor variables to be included in the \code{h} function. Each row represents an observation and each column represents an predictor.}

\item{X}{an \code{n-by-K} matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.}

\item{Znew}{matrix of new predictor values at which to predict new h, where each row represents a new observation. If set to NULL then will default to using the observed exposures Z}

\item{sel}{selects which iterations of the MCMC sampler to use for inference}
}
\value{
A list of mean, variance, entire mean matrix and variance array
}
\description{
#' Compute the posterior mean and variance of \code{h} at a new predictor values
Function to estimate the posterior mean and variance by obtaining the posterior mean and
variance at #' particular iterations and then using the iterated mean and variance formulas.
This function returns the entire mean matrix and variance array
needed for each of the future functions to obtain an unbiased estimate
of the SE used to create CI in plots.
}
