% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVcomps.R
\name{makeKpart}
\alias{makeKpart}
\title{makeKpart}
\usage{
makeKpart(r, Z1, Z2 = NULL)
}
\arguments{
\item{r}{need to update}

\item{Z1}{need to update}

\item{Z2}{need to update}
}
\value{
need to update
}
\description{
makeKpart
}
\examples{


}
