% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVcomps.R
\name{makeVcomps}
\alias{makeVcomps}
\title{Title}
\usage{
makeVcomps(r, lambda, Z, data.comps)
}
\arguments{
\item{r}{need to update}

\item{lambda}{need to update}

\item{Z}{need to update}

\item{data.comps}{need to update}
}
\value{
need to update
}
\description{
Title
}
