% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverallRiskSummaries.fixEY.R
\name{OverallRiskSummaries.fixEY}
\alias{OverallRiskSummaries.fixEY}
\title{Calculate overall risk summaries when fixing multiple effect modifiers at certain levels}
\usage{
OverallRiskSummaries.fixEY(
  list.fit.y.TE,
  qs = seq(0.25, 0.75, by = 0.05),
  q.fixed = 0.5,
  q.alwaysfixed = NULL,
  EY.alwaysfixed.name = NULL,
  sel = NULL,
  method = "approx"
)
}
\arguments{
\item{list.fit.y.TE}{The Total Effect BKMR model fit in a 'List' form.}

\item{qs}{vector of quantiles at which to calculate the overall risk summary}

\item{q.fixed}{a second quantile at which to compare the estimated {h} function}

\item{q.alwaysfixed}{the quantile values in the point which we want to keep fixed for all comparisons}

\item{EY.alwaysfixed.name}{names of all the effect modifiers that we want to fixed for all comparisons}

\item{sel}{selects which iterations of the MCMC sampler to use for inference}

\item{method}{method for obtaining posterior summaries at a vector of new points. Options are"approx" and "exact"; defaults to "approx", which is faster particularly for large datasets}
}
\value{
a list of data frame containing the (posterior mean) estimate and posterior standard deviation of the overall risk measures, for each of the comparisons specified
}
\description{
Calculate overall risk summaries when fixing multiple effect modifiers at certain levels
}
\details{
For guided examples, go to https://zc2326.github.io/causalbkmr/articles/BKMRCMA_Effectof_singleZ.html
}
