% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverallRiskSummaries.MI.R
\name{OverallRiskSummaries.MI}
\alias{OverallRiskSummaries.MI}
\title{Calculate overall risk summaries for BKMR MI}
\usage{
OverallRiskSummaries.MI(
  BKMRfits,
  qs = seq(0.25, 0.75, by = 0.05),
  q.fixed = 0.5,
  q.alwaysfixed = NULL,
  index.alwaysfixed = NULL,
  sel = NULL,
  method = "approx"
)
}
\arguments{
\item{BKMRfits}{A list of multiple BKMR fits and that each of these fits were ran for the same number of MCMC iterations.}

\item{qs}{vector of quantiles at which to calculate the overall risk summary}

\item{q.fixed}{a second quantile at which to compare the estimated {h} function}

\item{q.alwaysfixed}{the quantile values in the point which we want to keep fixed for all comparisons}

\item{index.alwaysfixed}{the index values in the point which we want to keep fixed for all comparisons}

\item{sel}{selects which iterations of the MCMC sampler to use for inference}

\item{method}{method for obtaining posterior summaries at a vector of new points. Options are"approx" and "exact"; defaults to "approx", which is faster particularly for large datasets}
}
\value{
a data frame containing the (posterior mean) estimate and posterior standard deviation of the overall risk measures
}
\description{
Compare estimated h function when all predictors are at a particular quantile to when all are at a second fixed quantile
}
\details{
For guided examples, go to https://zc2326.github.io/causalbkmr/articles/MI_BKMR.html
}
\examples{
library(causalbkmr)
data("BKMRfits10")
sel.MI <- seq(5001,10000, by = 5)

overallrisks.MI <- OverallRiskSummaries.MI(BKMRfits = BKMRfits10, qs = seq(0.1, 0.9, by = 0.05),  q.fixed = 0.5, sel = sel.MI, method = "approx")
ggplot(overallrisks.MI, aes(quantile, est, ymin = est - 1.96 * sd, ymax = est + 1.96 * sd)) +
geom_hline(yintercept = 00, linetype = "dashed", color = "gray") +
  geom_pointrange()+ ggtitle("") + scale_y_continuous(name="estimate")

}
