% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postresults.R
\name{postresults}
\alias{postresults}
\title{Posterior/Bootstrap Summary Function}
\usage{
postresults(posteriorsamp, alpha)
}
\arguments{
\item{posteriorsamp}{sample prediction}

\item{alpha}{1-confidence interval}
}
\value{
Mean. median , CI and sd of the sample prediction
}
\description{
Calculate mean/median/CI/sd of a vector of posterior/bootstrap samples
}
