% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rubin.MI.R
\name{Rubin.MI}
\alias{Rubin.MI}
\title{Rubins 1987 method}
\usage{
Rubin.MI(mean.vec, variance.vec)
}
\arguments{
\item{mean.vec}{A vector consist of mean values}

\item{variance.vec}{A vector consist of variance values}
}
\value{
Point estimates and std errors when combining information across MI fits (for approx method)
}
\description{
Function to implement Rubins 1987 method to estimate point estimates and std errors when combining information across MI fits (for approx method)
}
