% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingVarRiskSummaries.MI.R
\name{SingVarRiskSummaries.MI}
\alias{SingVarRiskSummaries.MI}
\title{Single Variable Risk Summaries for MI BKMR fits
Compute summaries of the risks associated with a change in a single variable in \code{Z} from a single level (quantile) to a second level (quantile), for the other variables in \code{Z} fixed to a specific level (quantile) for MI BKMR fits}
\usage{
SingVarRiskSummaries.MI(
  BKMRfits,
  which.z = 1:ncol(BKMRfits[[1]]$Z),
  qs.diff = c(0.25, 0.75),
  q.fixed = c(0.25, 0.5, 0.75),
  q.alwaysfixed = NULL,
  index.alwaysfixed = NULL,
  sel = NULL,
  z.names = colnames(BKMRfits[[1]]$Z),
  method = "approx",
  ...
)
}
\arguments{
\item{BKMRfits}{A list of multiple BKMR fits and that each of these fits were ran for the same number of MCMC iterations.}

\item{which.z}{vector indicating which variables (columns of Z) for which the summary should be computed}

\item{qs.diff}{vector indicating the two quantiles q_1 and q_2 at which to compute \code{h(z_{q2}) -h(z_{q1})}}

\item{q.fixed}{vector of quantiles at which to fix the remaining predictors in Z}

\item{q.alwaysfixed}{the quantile values in the point which we want to keep fixed for all comparisons}

\item{index.alwaysfixed}{the index values in the point which we want to keep fixed for all comparisons}

\item{sel}{selects which iterations of the MCMC sampler to use for inference}

\item{z.names}{column names of the selected columns of Z in which.z}

\item{method}{method for obtaining posterior summaries at a vector of new points. Options are"approx" and "exact"; defaults to "approx", which is faster particularly for large datasets}

\item{...}{other arguments to pass on to the prediction function}
}
\value{
a data frame containing the (posterior mean) estimate and posterior standard deviation of the  predictor risk measures
}
\description{
Single Variable Risk Summaries for MI BKMR fits
Compute summaries of the risks associated with a change in a single variable in \code{Z} from a single level (quantile) to a second level (quantile), for the other variables in \code{Z} fixed to a specific level (quantile) for MI BKMR fits
}
\examples{

}
