% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.bkmr.R
\name{TE.bkmr}
\alias{TE.bkmr}
\title{Estimate total effect for BKMR}
\usage{
TE.bkmr(a, astar, e.y, fit.y.TE, X.predict.Y, alpha = 0.05, sel, seed)
}
\arguments{
\item{a}{exposure variables at current level}

\item{astar}{exposure variables at counterfactual level}

\item{e.y}{effect modifier for the outcome variable}

\item{fit.y.TE}{total effect model fit regressing outcome on exposures, effect modifiers and confounders on outcome}

\item{X.predict.Y}{counfounders for outcome}

\item{alpha}{1-confidence interval}

\item{sel}{a vector selecting which iterations of the fit should be retained or inference}

\item{seed}{the random seed to use to evaluate the code}
}
\value{
Totak effect for BKMR
}
\description{
Estimate total effect for BKMR
}
