% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TERiskSummaries.CMA.R
\name{TERiskSummaries.CMA}
\alias{TERiskSummaries.CMA}
\title{TERiskSummaries.CMA}
\usage{
TERiskSummaries.CMA(
  fit.TE,
  e.y = NULL,
  e.y.names = NULL,
  qs = seq(0.25, 0.75, by = 0.05),
  q.fixed = 0.5,
  alpha = 0.05,
  sel,
  seed = 122
)
}
\arguments{
\item{fit.TE}{total effect model fit regressing outcome on exposures, effect modifiers and confounders on outcome}

\item{e.y}{effect modifier for the outcome variable}

\item{e.y.names}{column name of the effect modifier for the outcome variable}

\item{qs}{vector of quantiles at which to calculate the overall risk summary}

\item{q.fixed}{a second quantile at which to compare the estimated h function}

\item{alpha}{1-confidence interval}

\item{sel}{a vector selecting which iterations of the fit should be retained or inference}

\item{seed}{the random seed to use to evaluate the code}
}
\value{
a data frame containing the (posterior mean) estimate and posterior standard deviation of the total effect risk measures
}
\description{
Compare estimated Total Effect when all predictors are at a particular quantile to when all are at a second fixed quantile
}
\details{
For guided examples, go to
https://zc2326.github.io/causalbkmr/articles/BKMRCMA_QuickStart.html
}
\examples{
\dontrun{
library(causalbkmr)
riskSummary10 = TERiskSummaries.CMA(fit.TE = fit.y.TE, e.y=e.y10, e.y.name = "E.Y", sel=sel)

ggplot(riskSummary10,
       aes(quantile,
           est,
           ymin = est - 1.96 * sd,
           ymax = est + 1.96 * sd)) +
  geom_pointrange()
}

}
