% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarRiskSummary.CMA.R
\name{VarRiskSummary.CMA}
\alias{VarRiskSummary.CMA}
\title{VarRiskSummary.CMA}
\usage{
VarRiskSummary.CMA(
  whichz = 1,
  BKMRfits,
  e.y = NULL,
  e.y.names = NULL,
  qs.diff = c(0.25, 0.75),
  q.fixed = 0.5,
  alpha,
  sel,
  seed
)
}
\arguments{
\item{whichz}{vector indicating which variables (columns of Z) for which the summary should be computed}

\item{BKMRfits}{An object contatinint the results return by the kmbayes function}

\item{e.y}{effect modifier for the outcome variable}

\item{e.y.names}{column name of the effect modifier for the outcome variable}

\item{qs.diff}{vector indicating the two quantiles q_1 and q_2 at which to compute \code{h(z_{q2}) -h(z_{q1})}}

\item{q.fixed}{vector of quantiles at which to fix the remaining predictors in Z}

\item{alpha}{1-confidence interval}

\item{sel}{a vector selecting which iterations of the fit should be retained or inference}

\item{seed}{the random seed to use to evaluate the code}
}
\value{
a data frame containing the (posterior mean) estimate and posterior standard deviation of the  predictor risk measures
}
\description{
A combination function of VarRiskSummary and riskSummary.approx for MI BKMR fits, compare estimated \code{h} function when a single variable (or a set of variables) is at the 75th versus 25th percentile, when all of the other variables are fixed at a particular percentile
}
