% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YaMastar.SamplePred.R
\name{YaMastar.SamplePred}
\alias{YaMastar.SamplePred}
\title{YaMastar counterfactual}
\usage{
YaMastar.SamplePred(
  a,
  astar,
  e.y,
  fit.m,
  fit.y,
  X.predict.M,
  X.predict.Y,
  sel,
  seed,
  K
)
}
\arguments{
\item{a}{exposure variables at current level}

\item{astar}{exposure variables at counterfactual level}

\item{e.y}{effect modifier for the outcome variable}

\item{fit.m}{model fit regressing mediator on exposures and confounders on mediator}

\item{fit.y}{model fit regressing outcome on exposures, effect modifiers, mediator and confounders on outcome}

\item{X.predict.M}{counfounders for mediator}

\item{X.predict.Y}{counfounders for outcome}

\item{sel}{a vector selecting which iterations of the fit should be retained or inference}

\item{seed}{the random seed to use to evaluate the code}

\item{K}{number of samples to generate for each MCMC iteration}
}
\value{
A vector containing the sample prediction for YaMastar
}
\description{
Get YaMastar used in NDE and NIE caculation
}
