% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.bkmr.R
\name{mediation.bkmr}
\alias{mediation.bkmr}
\title{Estimate NDE/NIE for BKMR(plus TE)}
\usage{
mediation.bkmr(
  a,
  astar,
  e.y,
  fit.m = NULL,
  fit.y = NULL,
  fit.y.TE = NULL,
  X.predict.M = NULL,
  X.predict.Y = NULL,
  effects = "all",
  m.quant = c(0.1, 0.5, 0.75),
  m.value = NULL,
  alpha = 0.05,
  sel,
  seed,
  K
)
}
\arguments{
\item{a}{exposure variables at current level}

\item{astar}{exposure variables at counterfactual level}

\item{e.y}{effect modifier for the outcome variable}

\item{fit.m}{model fit regressing mediator on exposures and confounders on mediator}

\item{fit.y}{model fit regressing outcome on exposures, effect modifiers, mediator and confounders on outcome}

\item{fit.y.TE}{total effect model fit regressing outcome on exposures, effect modifiers and confounders on outcome}

\item{X.predict.M}{counfounders for mediator}

\item{X.predict.Y}{counfounders for outcome}

\item{effects}{type(s) of effects that users want to output}

\item{m.quant}{values of the quantile that the mediator is set to}

\item{m.value}{values that the mediator is set to}

\item{alpha}{1-confidence interval}

\item{sel}{a vector selecting which iterations of the fit should be retained or inference}

\item{seed}{the random seed to use to evaluate the code}

\item{K}{number of samples to generate for each MCMC iteration in YaMastar calculation}
}
\value{
A list contaning the sample prediction for TE, NDE, NIE and their summary statistics
}
\description{
Estimate NDE/NIE for BKMR(plus TE)
}
