% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingVarRiskSummaries.CMA.R
\name{SingVarRiskSummaries.CMA}
\alias{SingVarRiskSummaries.CMA}
\title{Single Variable Risk Summaries for CMA
Compute summaries of the risks associated with a change in a single variable in \code{Z} from a single level (quantile) to a second level (quantile), for the other variables in \code{Z} fixed to a specific level (quantile)}
\usage{
SingVarRiskSummaries.CMA(
  BKMRfits,
  e.y = NULL,
  e.y.names = NULL,
  which.z = 1:ncol(BKMRfits$Z),
  z.names = colnames(BKMRfits$Z),
  qs.diff = c(0.25, 0.75),
  q.fixed = c(0.25, 0.5, 0.75),
  alpha = 0.05,
  sel,
  seed = 122
)
}
\arguments{
\item{BKMRfits}{An object containing the results returned by a the kmbayes function}

\item{e.y}{effect modifier for the outcome variable}

\item{e.y.names}{column name of the effect modifier for the outcome variable}

\item{which.z}{vector indicating which variables (columns of Z) for which the summary should be computed}

\item{z.names}{optional vector of names for the columns of z}

\item{qs.diff}{vector indicating the two quantiles q_1 and q_2 at which to compute h(z_{q2}) -h(z_{q1})}

\item{q.fixed}{vector of quantiles at which to fix the remaining predictors in Z}

\item{alpha}{1-confidence interval}

\item{sel}{a vector selecting which iterations of the fit should be retained or inference}

\item{seed}{the random seed to use to evaluate the code}
}
\value{
a data frame containing the (posterior mean) estimate and posterior standard deviation of the single predictor risk measures
}
\description{
Single Variable Risk Summaries for CMA
Compute summaries of the risks associated with a change in a single variable in \code{Z} from a single level (quantile) to a second level (quantile), for the other variables in \code{Z} fixed to a specific level (quantile)
}
