% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.bkmr.R
\name{YaYastar.SamplePred}
\alias{YaYastar.SamplePred}
\title{Estimate TE for BKMR}
\usage{
YaYastar.SamplePred(a, astar, e.y, fit.y.TE, X.predict.Y, sel, seed)
}
\arguments{
\item{a}{exposure variables at current level}

\item{astar}{exposure variables at counterfactual level}

\item{e.y}{effect modifier for the outcome variable}

\item{fit.y.TE}{total effect model fit regressing outcome on exposures, effect modifiers and confounders on outcome}

\item{X.predict.Y}{counfounders for outcome}

\item{sel}{a vector selecting which iterations of the fit should be retained or inference}

\item{seed}{the random seed to use to evaluate the code}
}
\value{
A list containing the sample prediction for Ya and Yastar
}
\description{
Get Ya and Yastar used in effects caculation
}
